avoid various downcasting warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@20231 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/port/cpl_vsi_mem.cpp
@@ -211,7 +211,8 @@ bool VSIMemFile::SetLength( vsi_l_offset nNewLength )
             return false;
             
         /* Clear the new allocated part of the buffer */
-        memset(pabyNewData + nAllocLength, 0, nNewAlloc - nAllocLength);
+        memset(pabyNewData + nAllocLength, 0, 
+               (size_t) (nNewAlloc - nAllocLength));
 
         pabyData = pabyNewData;
         nAllocLength = nNewAlloc;
